+++ gip/operation/Interpolate.h
@@ -79,7 +79,7 @@ namespace gip {
   };
   
   template<>
-  class Interpolate<GrayPixel> : BinaryOperation<long double, long double, long double> {
+  class Interpolate<GrayPixel> : BinaryOperation<double, double, double> {
   private:
     
     typedef GrayPixel Pixel;
@@ -91,24 +91,24 @@ namespace gip {
       : elements(image.getElements()), dimension(image.getDimension()) {
     }
     
-    inline long double operator()(long double x, long double y) const noexcept {
+    inline double operator()(double x, double y) const noexcept {
       const int x0 = static_cast<int>(Math::floor(x));
       const int y0 = static_cast<int>(Math::floor(y));
       
-      const long double xFraction = x - x0;
-      const long double yFraction = y - y0;
+      const double xFraction = x - x0;
+      const double yFraction = y - y0;
       
-      long double result = 0; // 0 is background
+      double result = 0; // 0 is background
 
       const Pixel* p = elements + y0 * dimension.getWidth() + x0;
       
       if ((x0 >= 0) && (static_cast<unsigned int>(x0) < dimension.getWidth())) {
         if ((y0 >= 0) && (static_cast<unsigned int>(y0) < dimension.getHeight())) {
-          const long double w0 = (1 - xFraction) * (1 - yFraction);
+          const double w0 = (1 - xFraction) * (1 - yFraction);
           result += w0 * p[0];
         }
         if (((y0 + 1) >= 0) && (static_cast<unsigned int>(y0 + 1) < dimension.getHeight())) {
-          const long double w2 = (1 - xFraction) * yFraction;
+          const double w2 = (1 - xFraction) * yFraction;
           result += w2 * p[dimension.getWidth()];
         }
       }
@@ -117,11 +117,11 @@ namespace gip {
       
       if (((x0 + 1) >= 0) && (static_cast<unsigned int>(x0 + 1) < dimension.getWidth())) {
         if ((y0 >= 0) && (static_cast<unsigned int>(y0) < dimension.getHeight())) {
-          const long double w1 = xFraction * (1 - yFraction);
+          const double w1 = xFraction * (1 - yFraction);
           result += w1 * p[0];
         }
         if (((y0 + 1) >= 0) && (static_cast<unsigned int>(y0 + 1) < dimension.getHeight())) {
-          const long double w3 = xFraction * yFraction;
+          const double w3 = xFraction * yFraction;
           result += w3 * p[dimension.getWidth()];
         }
       }
@@ -131,7 +131,7 @@ namespace gip {
   };
   
   template<class COMPONENT>
-  class Interpolate<RGBPixel<COMPONENT> > : BinaryOperation<RGBPixel<long double>, long double, long double> {
+  class Interpolate<RGBPixel<COMPONENT> > : BinaryOperation<RGBPixel<double>, double, double> {
   private:
     
     typedef RGBPixel<COMPONENT> Pixel;
@@ -143,28 +143,28 @@ namespace gip {
       : elements(image.getElements()), dimension(image.getDimension()) {
     }
     
-    inline RGBPixel<COMPONENT> operator()(long double x, long double y) const noexcept {
+    inline RGBPixel<COMPONENT> operator()(double x, double y) const noexcept {
       const int x0 = static_cast<int>(Math::floor(x));
       const int y0 = static_cast<int>(Math::floor(y));
       
-      const long double xFraction = x - x0;
-      const long double yFraction = y - y0;
+      const double xFraction = x - x0;
+      const double yFraction = y - y0;
       
-      long double red = 0; // 0 is background
-      long double green = 0;
-      long double blue = 0;
+      double red = 0; // 0 is background
+      double green = 0;
+      double blue = 0;
       
       const Pixel* p = elements + y0 * dimension.getWidth() + x0;
       
       if ((x0 >= 0) && (static_cast<unsigned int>(x0) < dimension.getWidth())) {
         if ((y0 >= 0) && (static_cast<unsigned int>(y0) < dimension.getHeight())) {
-          const long double w0 = (1 - xFraction) * (1 - yFraction);
+          const double w0 = (1 - xFraction) * (1 - yFraction);
           red += w0 * p[0].red;
           green += w0 * p[0].green;
           blue += w0 * p[0].blue;
         }
         if (((y0 + 1) >= 0) && (static_cast<unsigned int>(y0 + 1) < dimension.getHeight())) {
-          const long double w2 = (1 - xFraction) * yFraction;
+          const double w2 = (1 - xFraction) * yFraction;
           red += w2 * p[dimension.getWidth()].red;
           green += w2 * p[dimension.getWidth()].green;
           blue += w2 * p[dimension.getWidth()].blue;
@@ -175,13 +175,13 @@ namespace gip {
       
       if (((x0 + 1) >= 0) && (static_cast<unsigned int>(x0 + 1) < dimension.getWidth())) {
         if ((y0 >= 0) && (static_cast<unsigned int>(y0) < dimension.getHeight())) {
-          const long double w1 = xFraction * (1 - yFraction);
+          const double w1 = xFraction * (1 - yFraction);
           red += w1 * p[0].red;
           green += w1 * p[0].green;
           blue += w1 * p[0].blue;
         }
         if (((y0 + 1) >= 0) && (static_cast<unsigned int>(y0 + 1) < dimension.getHeight())) {
-          const long double w3 = xFraction * yFraction;
+          const double w3 = xFraction * yFraction;
           red += w3 * p[dimension.getWidth()].red;
           green += w3 * p[dimension.getWidth()].green;
           blue += w3 * p[dimension.getWidth()].blue;
